+++ modules/apps/collaboration/message-boards/message-boards-comment/src/main/java/com/liferay/message/boards/comment/internal/MBDiscussionStagingHandler.java
@@ -34,7 +34,9 @@ import com.liferay.portal.kernel.dao.orm.ExportActionableDynamicQuery;
 import com.liferay.portal.kernel.dao.orm.Property;
 import com.liferay.portal.kernel.dao.orm.PropertyFactoryUtil;
 import com.liferay.portal.kernel.dao.orm.RestrictionsFactoryUtil;
+import com.liferay.portal.kernel.model.Group;
 import com.liferay.portal.kernel.model.StagedModel;
+import com.liferay.portal.kernel.service.GroupLocalServiceUtil;
 import com.liferay.portal.kernel.workflow.WorkflowConstants;
 import com.liferay.portlet.messageboards.service.permission.MBPermission;
 
@@ -56,6 +58,13 @@ public class MBDiscussionStagingHandler implements DiscussionStagingHandler {
 				ExportImportClassedModelUtil.getClassPK(stagedModel));
 
 		if (mbDiscussion == null) {
+			return;
+		}
+
+		Group group = GroupLocalServiceUtil.fetchGroup(
+			mbDiscussion.getGroupId());
+
+		if ((group != null) && group.isControlPanel()) {
 			return;
 		}