+++ modules/apps/collaboration/wiki/wiki-web/src/main/java/com/liferay/wiki/web/internal/portlet/action/ExportPageMVCActionCommand.java
@@ -77,6 +77,8 @@ public class ExportPageMVCActionCommand extends BaseMVCActionCommand {
 			ActionRequest actionRequest, ActionResponse actionResponse)
 		throws Exception {
 
+		hideDefaultSuccessMessage(actionRequest);
+
 		PortletConfig portletConfig = getPortletConfig(actionRequest);
 
 		try {