+++ modules/apps/foundation/portal-language/portal-language-servlet-filter/src/main/java/com/liferay/portal/language/servlet/filter/internal/LanguageFilterTracker.java
@@ -19,6 +19,7 @@ import com.liferay.portal.kernel.util.AggregateResourceBundle;
 import com.liferay.portal.kernel.util.LocaleUtil;
 import com.liferay.portal.kernel.util.ResourceBundleLoader;
 import com.liferay.portal.kernel.util.ResourceBundleLoaderUtil;
+import com.liferay.portal.kernel.util.StringBundler;
 
 import java.util.ArrayList;
 import java.util.Dictionary;
@@ -130,24 +131,52 @@ public class LanguageFilterTracker {
 
 			Bundle bundle = serviceReference.getBundle();
 
-			String filterString =
-				"(&(bundle.symbolic.name=" + bundle.getSymbolicName() +
-					")(objectClass=" + ResourceBundleLoader.class.getName() +
-						")(resource.bundle.base.name=*))";
+			StringBundler filterSB = new StringBundler();
+
+			Object contextName = serviceReference.getProperty(
+				HttpWhiteboardConstants.HTTP_WHITEBOARD_CONTEXT_NAME);
+
+			if (contextName == null) {
+				filterSB.append("(&");
+				filterSB.append("(bundle.symbolic.name=");
+				filterSB.append(bundle.getSymbolicName());
+				filterSB.append(")");
+				filterSB.append("(objectClass=");
+				filterSB.append(ResourceBundleLoader.class.getName());
+				filterSB.append(")");
+				filterSB.append("(resource.bundle.base.name=*)");
+				filterSB.append(")");
+			}
+			else {
+				filterSB.append("(&");
+				filterSB.append("(|");
+				filterSB.append("(bundle.symbolic.name=");
+				filterSB.append(bundle.getSymbolicName());
+				filterSB.append(")");
+				filterSB.append("(&");
+				filterSB.append("(servlet.context.name=");
+				filterSB.append(contextName);
+				filterSB.append(")");
+				filterSB.append("(service.bundleid=0)");
+				filterSB.append(")");
+				filterSB.append(")");
+				filterSB.append("(objectClass=");
+				filterSB.append(ResourceBundleLoader.class.getName());
+				filterSB.append(")");
+				filterSB.append("(resource.bundle.base.name=*)");
+				filterSB.append(")");
+			}
 
 			Map<String, Object> properties = new HashMap<>();
 
 			properties.put("service.ranking", Integer.MIN_VALUE);
 
-			Object contextName = serviceReference.getProperty(
-				HttpWhiteboardConstants.HTTP_WHITEBOARD_CONTEXT_NAME);
-
 			properties.put("servlet.context.name", contextName);
 
 			return ServiceTrackerFactory.open(
-				bundle.getBundleContext(), filterString,
+				bundle.getBundleContext(), filterSB.toString(),
 				new ResourceBundleLoaderServiceTrackerCustomizer(
-					properties, filterString, contextName));
+					properties, filterSB.toString(), contextName));
 		}
 
 		@Override