+++ modules/apps/static/portal-lpkg-deployer/portal-lpkg-deployer/src/main/java/com/liferay/portal/lpkg/deployer/internal/LPKGBundleTrackerCustomizer.java
@@ -166,15 +166,17 @@ public class LPKGBundleTrackerCustomizer
 				while (enumeration.hasMoreElements()) {
 					URL url = enumeration.nextElement();
 
-					if (_checkOverridden(symbolicName, url)) {
+					String location =
+						LPKGInnerBundleLocationUtil.generateInnerBundleLocation(
+							bundle, url.getPath());
+
+					if (_checkOverridden(symbolicName, url, location)) {
 						continue;
 					}
 
 					if (_isBundleInstalled(bundle, url)) {
 						continue;
 					}
-
-					String location = url.getPath();
 
 					Bundle newBundle = _bundleContext.getBundle(location);
 
@@ -186,6 +188,10 @@ public class LPKGBundleTrackerCustomizer
 
 					newBundle = _bundleContext.installBundle(
 						location, url.openStream());
+
+					if (newBundle.getState() == Bundle.UNINSTALLED) {
+						continue;
+					}
 
 					BundleStartLevelUtil.setStartLevelAndStart(
 						newBundle,
@@ -206,11 +212,13 @@ public class LPKGBundleTrackerCustomizer
 			while (enumeration.hasMoreElements()) {
 				URL url = enumeration.nextElement();
 
-				if (_checkOverridden(symbolicName, url)) {
+				String location =
+					LPKGInnerBundleLocationUtil.generateInnerBundleLocation(
+						bundle, url.getPath());
+
+				if (_checkOverridden(symbolicName, url, location)) {
 					continue;
 				}
-
-				String location = url.getPath();
 
 				Bundle newBundle = _bundleContext.getBundle(location);
 
@@ -227,7 +235,11 @@ public class LPKGBundleTrackerCustomizer
 				// uninstalled, its wrapped WAR bundle will also be unintalled.
 
 				newBundle = _bundleContext.installBundle(
-					url.getPath(), _toWARWrapperBundle(bundle, url));
+					location, _toWARWrapperBundle(bundle, url));
+
+				if (newBundle.getState() == Bundle.UNINSTALLED) {
+					continue;
+				}
 
 				BundleStartLevelUtil.setStartLevelAndStart(
 					newBundle,
@@ -373,7 +385,8 @@ public class LPKGBundleTrackerCustomizer
 		return sb.toString();
 	}
 
-	private boolean _checkOverridden(String symbolicName, URL url)
+	private boolean _checkOverridden(
+			String symbolicName, URL url, String location)
 		throws Throwable {
 
 		String path = url.getPath();
@@ -387,7 +400,7 @@ public class LPKGBundleTrackerCustomizer
 		path = StringUtil.toLowerCase(path);
 
 		if (_overrideFileNames.contains(path)) {
-			Bundle bundle = _bundleContext.getBundle(url.getPath());
+			Bundle bundle = _bundleContext.getBundle(location);
 
 			if (bundle != null) {
 				_uninstallBundle(symbolicName.concat(StringPool.DASH), bundle);
@@ -419,7 +432,9 @@ public class LPKGBundleTrackerCustomizer
 			Version version = new Version(
 				attributes.getValue(Constants.BUNDLE_VERSION));
 
-			String location = url.getPath();
+			String location =
+				LPKGInnerBundleLocationUtil.generateInnerBundleLocation(
+					bundle, url.getPath());
 
 			for (Bundle installedBundle : _bundleContext.getBundles()) {
 				if (symbolicName.equals(installedBundle.getSymbolicName()) &&
@@ -587,6 +602,10 @@ public class LPKGBundleTrackerCustomizer
 
 	private void _uninstallBundle(String prefix, Bundle bundle)
 		throws Throwable {
+
+		if (bundle.getState() == Bundle.UNINSTALLED) {
+			return;
+		}
 
 		String symbolicName = bundle.getSymbolicName();