+++ modules/util/portal-tools-db-upgrade-client/src/main/java/com/liferay/portal/tools/db/upgrade/client/AppServer.java
@@ -14,9 +14,8 @@
 
 package com.liferay.portal.tools.db.upgrade.client;
 
-import com.liferay.portal.tools.db.upgrade.client.util.StringUtil;
-
 import java.io.File;
+import java.io.IOException;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -80,12 +79,12 @@ public class AppServer {
 		String dirName, String extraLibDirNames, String globalLibDirName,
 		String portalDirName, String serverDetectorServerId) {
 
-		_dir = new File(dirName);
-		_globalLibDir = new File(dirName, globalLibDirName);
-		_portalDir = new File(dirName, portalDirName);
-		_serverDetectorServerId = serverDetectorServerId;
+		_setDirName(dirName);
 
-		_setExtraLibDirNames(extraLibDirNames);
+		_extraLibDirNames = extraLibDirNames;
+		_globalLibDirName = globalLibDirName;
+		_portalDirName = portalDirName;
+		_serverDetectorServerId = serverDetectorServerId;
 	}
 
 	public File getDir() {
@@ -93,27 +92,43 @@ public class AppServer {
 	}
 
 	public String getExtraLibDirNames() {
-		return StringUtil.join(_extraLibDirs, ',');
+		return _extraLibDirNames;
 	}
 
 	public List<File> getExtraLibDirs() {
-		return _extraLibDirs;
+		List<File> extraLibDirs = new ArrayList<>();
+
+		if ((_extraLibDirNames != null) && !_extraLibDirNames.isEmpty()) {
+			for (String extraLibDirName : _extraLibDirNames.split(",")) {
+				extraLibDirs.add(new File(_dir, extraLibDirName));
+			}
+		}
+
+		return extraLibDirs;
 	}
 
 	public File getGlobalLibDir() {
-		return _globalLibDir;
+		return new File(_dir, _globalLibDirName);
+	}
+
+	public String getGlobalLibDirName() {
+		return _globalLibDirName;
 	}
 
 	public File getPortalClassesDir() {
-		return new File(_portalDir, "/WEB-INF/classes");
+		return new File(getPortalDir(), "/WEB-INF/classes");
 	}
 
 	public File getPortalDir() {
-		return _portalDir;
+		return new File(_dir, _portalDirName);
+	}
+
+	public String getPortalDirName() {
+		return _portalDirName;
 	}
 
 	public File getPortalLibDir() {
-		return new File(_portalDir, "/WEB-INF/lib");
+		return new File(getPortalDir(), "/WEB-INF/lib");
 	}
 
 	public String getServerDetectorServerId() {
@@ -121,19 +136,19 @@ public class AppServer {
 	}
 
 	public void setDirName(String dirName) {
-		_dir = new File(dirName);
+		_setDirName(dirName);
 	}
 
 	public void setExtraLibDirNames(String extraLibDirNames) {
-		_setExtraLibDirNames(extraLibDirNames);
+		_extraLibDirNames = extraLibDirNames;
 	}
 
 	public void setGlobalLibDirName(String globalLibDirName) {
-		_globalLibDir = new File(_dir, globalLibDirName);
+		_globalLibDirName = globalLibDirName;
 	}
 
 	public void setPortalDirName(String portalDirName) {
-		_portalDir = new File(_dir, portalDirName);
+		_portalDirName = portalDirName;
 	}
 
 	private static String _getJBossExtraLibDirNames() {
@@ -154,18 +169,23 @@ public class AppServer {
 		return sb.toString();
 	}
 
-	private void _setExtraLibDirNames(String extraLibDirNames) {
-		if ((extraLibDirNames != null) && !extraLibDirNames.isEmpty()) {
-			for (String extraLibDirName : extraLibDirNames.split(",")) {
-				_extraLibDirs.add(new File(_dir, extraLibDirName));
+	private void _setDirName(String dirName) {
+		try {
+			_dir = new File(dirName);
+
+			if (!_dir.isAbsolute()) {
+				_dir = _dir.getCanonicalFile();
 			}
+		}
+		catch (IOException ioe) {
+			ioe.printStackTrace();
 		}
 	}
 
 	private File _dir;
-	private final List<File> _extraLibDirs = new ArrayList<>();
-	private File _globalLibDir;
-	private File _portalDir;
+	private String _extraLibDirNames;
+	private String _globalLibDirName;
+	private String _portalDirName;
 	private final String _serverDetectorServerId;
 
 }