+++ portal-impl/src/com/liferay/portlet/asset/action/GetCategoriesAction.java
@@ -72,6 +72,7 @@ public class GetCategoriesAction extends JSONAction {
 	protected List<AssetCategory> getCategories(HttpServletRequest request)
 		throws Exception {
 
+		long scopeGroupId = ParamUtil.getLong(request, "scopeGroupId");
 		long categoryId = ParamUtil.getLong(request, "categoryId");
 		long vocabularyId = ParamUtil.getLong(request, "vocabularyId");
 		int start = ParamUtil.getInteger(request, "start", QueryUtil.ALL_POS);
@@ -80,16 +81,29 @@ public class GetCategoriesAction extends JSONAction {
 		List<AssetCategory> categories = Collections.emptyList();
 
 		if (categoryId > 0) {
-			categories = AssetCategoryServiceUtil.getChildCategories(
-				categoryId, start, end, null);
+			if (scopeGroupId > 0) {
+				categories = AssetCategoryServiceUtil.getVocabularyCategories(
+					scopeGroupId, categoryId, vocabularyId, start, end, null);
+			}
+			else {
+				categories = AssetCategoryServiceUtil.getChildCategories(
+					categoryId, start, end, null);
+			}
 		}
 		else if (vocabularyId > 0) {
 			long parentCategoryId = ParamUtil.getLong(
 				request, "parentCategoryId",
 				AssetCategoryConstants.DEFAULT_PARENT_CATEGORY_ID);
 
-			categories = AssetCategoryServiceUtil.getVocabularyCategories(
-				parentCategoryId, vocabularyId, start, end, null);
+			if (scopeGroupId > 0) {
+				categories = AssetCategoryServiceUtil.getVocabularyCategories(
+					scopeGroupId, parentCategoryId, vocabularyId, start, end,
+					null);
+			}
+			else {
+				categories = AssetCategoryServiceUtil.getVocabularyCategories(
+					parentCategoryId, vocabularyId, start, end, null);
+			}
 		}
 
 		return categories;